requirements.txt
nker==1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudinary==1.23.0
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-PyMongo==2.3.0
freeze==3.0
gunicorn==20.0.4
heroku==0.1.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
langdetect==1.0.8
MarkupSafe==1.1.1
numpy==1.19.5
opencv-python-headless==4.5.1.48
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.2
pytesseract==0.3.7
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
wincertstore==0.2
